Move Qwen KV trace helpers out of main
